Create Move details .jsx and moved empty the old one.
